+++ CHANGES.md
@@ -78,6 +78,7 @@
 * Add provider Anizb
 * Change TorrentDay to use its 2.x interface
 * Add button 'Discover' Emby server to notifications
+* Add Bit-HDTV torrent provider
 
 
 ### 0.11.11 (2016-04-05 19:20:00 UTC)